requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
azure-storage-blob==12.25.1
azure-core==1.34.0
azure-identity==1.23.0
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
attrs==24.3.0
beautifulsoup4==4.12.3
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dataclasses-json==0.6.7
decorator==5.2.1
Deprecated==1.2.15
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.7.0
docx2txt==0.8
executing==2.1.0
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.6
filetype==1.2.0
frozenlist==1.5.0
fsspec==2024.12.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
icecream==2.1.4
idna==3.10
ipython==8.37.0
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.8.2
joblib==1.4.2
jsonpickle==4.1.0
llama-cloud==0.1.8
llama-index==0.12.10
llama-index-agent-openai==0.4.1
llama-index-cli==0.4.0
llama-index-core==0.12.10.post1
llama-index-embeddings-azure-openai==0.3.9
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.3
llama-index-llms-azure-openai==0.3.2
llama-index-llms-ollama==0.5.0
llama-index-llms-openai==0.3.13
llama-index-multi-modal-llms-openai==0.4.2
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.3
llama-index-readers-llama-parse==0.4.0
llama-index-utils-workflow==0.3.2
llama-parse==0.5.19
MarkupSafe==3.0.2
marshmallow==3.25.1
matplotlib-inline==0.1.7
multidict==6.1.0
mypy-extensions==1.0.0
motor==3.7.1
msal==1.32.3
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.1
ollama==0.4.7
openai==1.59.7
packaging==24.2
pandas==2.2.3
parso==0.8.4
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
pexpect==4.9.0
prompt_toolkit==3.0.51
propcache==0.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pydantic[email]==2.10.5
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.13.0
PyMuPDF==1.25.3
pymupdf4llm==0.0.17
pypdf==5.1.0
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
pyvis==0.3.2
PyYAML==6.0.2
regex==2024.11.6
redis==6.2.0
requests==2.32.3
sendgrid==6.11.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
stack-data==0.6.3
starlette==0.41.3
striprtf==0.0.26
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
whisper==1.1.10
wrapt==1.17.2
yarl==1.18.3
